Add Telegram Bot
Create a Telegram bot by following the instructions given on the Telegram website. This code defines a get_weather function that takes a location as input, retrieves the current weather information for that location from the OpenWeatherMap API, and sends a message to the user with the weather information.
Users can interact with it by sending a message with the /weather command followed by a location. The bot will retrieve the current weather information for that location and send it back to the user.